requirements.txt
colorama
requests
urllib3
dnspython
html_similarity
socket
argparse